Move debug tests
